+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:qiita_client_yukiy/pages/top_page.dart';
 
 void main() {
   runApp(const MyApp());
@@ -10,107 +11,11 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Flutter Demo',
+      title: 'Qiita client yukiy',
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: const MyHomePage(title: 'Flutter Demo Home Page'),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({Key? key, required this.title}) : super(key: key);
-
-  final String title;
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Stack(
-        children: <Widget>[
-          Container(
-            decoration: BoxDecoration(
-              color: Colors.black.withOpacity(0.1),
-              image: const DecorationImage(
-                image: AssetImage('assets/images/background.png'),
-                fit: BoxFit.fill,
-              ),
-            ),
-          ),
-          Container(
-            color: Colors.black.withOpacity(0.2),
-          ),
-          Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: <Widget>[
-                const SizedBox(
-                  height: 100,
-                ),
-                Container(
-                  alignment: Alignment.center,
-                  child: const Text(
-                    'Qiita Feed App',
-                    style: TextStyle(
-                      color: Colors.white,
-                      fontSize: 36,
-                      fontFamily: 'Pacifico-Regular',
-                    ),
-                  ),
-                ),
-                Container(
-                  alignment: Alignment.center,
-                  child: const Text(
-                    '-PlayGround-',
-                    style: TextStyle(
-                      letterSpacing: 0.25,
-                      color: Colors.white,
-                      fontSize: 14,
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                ),
-                const Spacer(),
-                Container(
-                  height: 50,
-                  margin: const EdgeInsets.symmetric(horizontal: 24),
-                  child: ElevatedButton(
-                    onPressed: () {},
-                    style: ElevatedButton.styleFrom(
-                        textStyle: const TextStyle(
-                          fontSize: 14,
-                          fontWeight: FontWeight.bold,
-                        ),
-                        primary: const Color.fromRGBO(70, 131, 0, 1),
-                        shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(25),
-                        )),
-                    child: const Text('ログイン'),
-                  ),
-                ),
-                Container(
-                  margin: const EdgeInsets.all(34),
-                  width: 327,
-                  height: 50,
-                  child: TextButton(
-                    onPressed: () {},
-                    style: TextButton.styleFrom(
-                        foregroundColor: Colors.white,
-                        textStyle: const TextStyle(
-                          fontWeight: FontWeight.bold,
-                        )),
-                    child: const Text('ログインせずに利用する'),
-                  ),
-                ),
-              ])
-        ],
-      ),
+      home: const TopPage(),
     );
   }
 }
